FILE: renames internal folders to match new class names
